+++ demo/index.ts
@@ -658,7 +658,12 @@ class M1 {
   }
 }
 
-const inDiv = new InDiv();
-inDiv.bootstrapModule(M1);
-inDiv.use(PlatformBrowser);
-inDiv.init();
+// const inDiv = new InDiv();
+// inDiv.bootstrapModule(M1);
+// inDiv.use(PlatformBrowser);
+// inDiv.init();
+InDiv.bootstrap(M1, {
+  plugins: [
+    PlatformBrowser,
+  ],
+});