+++ ssr-demo/public/main.ts
@@ -6,9 +6,15 @@ import { InDiv } from '@indiv/core';
 import { PlatformBrowser } from '@indiv/platform-browser';
 import RootModule from './modules';
 
-export const inDiv = new InDiv();
-inDiv.bootstrapModule(RootModule);
-inDiv.use(PlatformBrowser);
-inDiv.init();
+// export const inDiv = new InDiv();
+// inDiv.bootstrapModule(RootModule);
+// inDiv.use(PlatformBrowser);
+// inDiv.init();
 
-console.log('indiv', inDiv);
+InDiv.bootstrap(RootModule, {
+  plugins: [
+    PlatformBrowser,
+  ],
+});
+
+// console.log('indiv', inDiv);